+++ relay-statsd/src/lib.rs
@@ -23,7 +23,7 @@
 //! ```no_run
 //! # use std::collections::BTreeMap;
 //!
-//! relay_statsd::init("myprefix", "localhost:8125", BTreeMap::new(), true, 1.0);
+//! relay_statsd::init("myprefix", "localhost:8125", BTreeMap::new(), true, true, 1.0);
 //! ```
 //!
 //! ## Macro Usage
@@ -66,6 +66,8 @@ use cadence::{
 };
 use parking_lot::RwLock;
 use rand::distributions::{Distribution, Uniform};
+use statsdproxy::cadence::StatsdProxyMetricSink;
+use statsdproxy::config::AggregateMetricsConfig;
 
 /// Maximum number of metric events that can be queued before we start dropping them
 const METRICS_MAX_QUEUE_SIZE: usize = 100_000;
@@ -226,6 +228,7 @@ pub fn init<A: ToSocketAddrs>(
     host: A,
     default_tags: BTreeMap<String, String>,
     buffering: bool,
+    aggregating: bool,
     sample_rate: f32,
 ) {
     let addrs: Vec<_> = host.to_socket_addrs().unwrap().collect();
@@ -247,7 +250,24 @@ pub fn init<A: ToSocketAddrs>(
     let socket = UdpSocket::bind("0.0.0.0:0").unwrap();
     socket.set_nonblocking(true).unwrap();
 
-    let statsd_client = if buffering {
+    let statsd_client = if aggregating {
+        let host = host.to_socket_addrs().unwrap().next().unwrap();
+        let statsdproxy_sink = StatsdProxyMetricSink::new(move || {
+            let next_step = statsdproxy::middleware::upstream::Upstream::new(host)
+                .expect("failed to create statsdproxy metric sink");
+            statsdproxy::middleware::aggregate::AggregateMetrics::new(
+                AggregateMetricsConfig {
+                    aggregate_gauges: true,
+                    aggregate_counters: true,
+                    flush_interval: 1,
+                    flush_offset: 0,
+                    max_map_size: None,
+                },
+                next_step,
+            )
+        });
+        StatsdClient::from_sink(prefix, statsdproxy_sink)
+    } else if buffering {
         let udp_sink = BufferedUdpMetricSink::from(host, socket).unwrap();
         let queuing_sink = QueuingMetricSink::with_capacity(udp_sink, METRICS_MAX_QUEUE_SIZE);
         StatsdClient::from_sink(prefix, queuing_sink)