+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Extend GPU context with data for Unreal Engine crash reports. ([#3144](https://github.com/getsentry/relay/pull/3144))
 - Parametrize transaction in dynamic sampling context. ([#3141](https://github.com/getsentry/relay/pull/3141))
 - Adds ReplayVideo envelope-item type. ([#3105](https://github.com/getsentry/relay/pull/3105))
+- Parse & scrub span description for supabase. ([#3153](https://github.com/getsentry/relay/pull/3153))
 
 **Bug Fixes**:
 
@@ -17,6 +18,7 @@
 
 - Add quotas to global config. ([#3086](https://github.com/getsentry/relay/pull/3086))
 - Adds support for dynamic metric bucket encoding. ([#3137](https://github.com/getsentry/relay/pull/3137))
+- Use statsdproxy to pre-aggregate metrics. ([#2425](https://github.com/getsentry/relay/pull/2425))
 
 ## 24.2.0